services/classhub/requirements.txt
Django==5.0.8
psycopg[binary]==3.2.1
django-environ==0.11.2
django-otp==1.6.3
whitenoise==6.7.0
gunicorn==22.0.0
redis==5.0.8

# Content packs (course.yaml + markdown lesson rendering)
PyYAML==6.0.2
Markdown==3.6
bleach==6.1.0
